third_party/WebKit/LayoutTests/svg/custom/getStrokeBBox.html

<!DOCTYPE html>
<html>
<head>
<style>
  body {
    margin: 0;
  }
  .measure {
    opacity: 0.7;
    stroke-width: 10;
    fill: lightgreen;
    stroke: lightblue;
  }
  #nostroke {
    stroke-width: 0;
  }
  .bbox {
    stroke-width: 1;
    fill: transparent;
    stroke: green;
  }
  .strokebbox {
    stroke-width: 1;
    fill: transparent;
    stroke: blue;
  }
</style>
<script>
  function runTest() {
    if (window.testRunner)
      testRunner.dumpAsText();
    printAndShowBoundingBoxes(['rect', 'ellipse', 'path', 'marker', 'capjoin', 'dash', 'nostroke']);
  }

  function printAndShowBoundingBoxes(ids) {
    var svg = document.getElementById('svg');
    var svgNS = "http://www.w3.org/2000/svg";

    ids.forEach(function(id) {
      var element = document.getElementById(id);
      var bbox = element.getBBox();
      var strokeBBox = element.getStrokeBBox()

      // Add a green rect for the bounding box.
      var bboxEl = document.createElementNS(svgNS, 'rect');
      bboxEl.setAttribute('x', bbox.x);
      bboxEl.setAttribute('y', bbox.y);
      bboxEl.setAttribute('width', bbox.width);
      bboxEl.setAttribute('height', bbox.height);
      bboxEl.setAttribute('class', 'bbox');
      svg.appendChild(bboxEl);

      // Add a blue rect for the stroke bounding box.
      var strokeBBoxEl = document.createElementNS(svgNS, 'rect');
      strokeBBoxEl.setAttribute('x', strokeBBox.x);
      strokeBBoxEl.setAttribute('y', strokeBBox.y);
      strokeBBoxEl.setAttribute('width', strokeBBox.width);
      strokeBBoxEl.setAttribute('height', strokeBBox.height);
      strokeBBoxEl.setAttribute('class', 'strokebbox');
      svg.appendChild(strokeBBoxEl);

      // Add text describing the bounding box and stroke bounding box.
      var bboxTextEl = document.createElementNS(svgNS, 'text');
      bboxTextEl.setAttribute('x', 10 + strokeBBox.x + strokeBBox.width);
      bboxTextEl.setAttribute('y', strokeBBox.y + strokeBBox.height);
      var bboxString =
          id + '.getBBox()=' + pretty(bbox.x) + ',' + pretty(bbox.y) + ',' + pretty(bbox.width) + ',' + pretty(bbox.height);
      var strokeBBoxString =
          id + '.getStrokeBBox()=' + pretty(strokeBBox.x) + ',' + pretty(strokeBBox.y) + ',' + pretty(strokeBBox.width) + ',' + pretty(strokeBBox.height);
      bboxTextEl.textContent = bboxString + ', ' + strokeBBoxString;
      svg.appendChild(bboxTextEl);
    });
  }

  function pretty(num) {
    return Math.round(num * 10) / 10;
  }
</script>
</head>
<body onload="runTest()">
  getStrokeBBox() box test for crbug.com/280423.<br/>
  <svg id="svg" width="600" height="500">
    <defs>
      <marker id="circle" refX="2" refY="2" markerWidth="4" markerHeight="4">
        <circle cx="2" cy="2" r="1.4"/>
      </marker>
    </defs>
    <rect id="rect" class="measure" x="20" y="20" width="40" height="40"/>
    <ellipse id="ellipse" class="measure" cx="40" cy="100" rx="10" ry="20"/>
    <path id="path" class="measure" d="M20 150 L 60 190 L 20 190Z"/>
    <path id="marker" class="measure" d="M20 220 L 60 260 L 20 260Z" marker-mid="url(#circle)" marker-start="url(#circle)"/>
    <path id="capjoin" class="measure" d="M20 300 L 60 340 L 20 340" stroke-linejoin="bevel" stroke-linecap="square"/>
    <path id="dash" class="measure" d="M20 360 L 60 400 L 20 400Z" stroke-dasharray="5 70" stroke-linecap="round"/>
    <rect id="nostroke" class="measure" x="20" y="420" width="40" height="40"/>
  </svg>
</body>
</html>
